bluetooth/getDevices/reject_opaque_origin.https.html

<!DOCTYPE html>
<head>
  <meta http-equiv="content-security-policy" content="sandbox allow-scripts"/>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <script src='/resources/testdriver.js'></script>
  <script src='/resources/testdriver-vendor.js'></script>
  <script src='/bluetooth/resources/bluetooth-test.js'></script>
  </head>
<html>
  <body>
    <script>
      bluetooth_test(async t => {
        promise_rejects_dom(
            t, 'SecurityError', navigator.bluetooth.getDevices(),
            'getDevices() should throw a SecutityError DOMException when called from a context where the top-level document has an opaque origin.');
      }, 'Calls to Bluetooth APIs from an origin with opaque top origin get blocked.');
    </script>
  </body>
</html>
